Refactor: 편의를 위해 percent 자료형을 long으로 변환한다.

+++ src/main/java/org/prgms/springbootbasic/domain/PercentDiscountVoucher.java
@@ -4,9 +4,9 @@ import java.util.UUID;
 
 public class PercentDiscountVoucher implements Voucher {
     private final UUID voucherId;
-    private final int percent;
+    private final long percent;
 
-    public PercentDiscountVoucher(UUID voucherId, int percent) {
+    public PercentDiscountVoucher(UUID voucherId, long percent) {
         if (percent > 100 || percent < 0)
             throw new IllegalArgumentException("percent value is out of range.");
